tests/wpt/tests/sanitizer-api/sanitizer-removeUnsafe.tentative.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(t => {
  // Create an empty config.
  let s = new Sanitizer({});
  // Remove everything unsafe.
  s.removeUnsafe();

  let config = s.get();
  assert_false('elements' in config, "no elements");
  assert_false('replaceWithChildrenElements' in config, "no replaceWithChildrenElements");
  assert_false('attributes' in config, "no attributes");

  // https://wicg.github.io/sanitizer-api/#built-in-safe-baseline-configuration
  const SAFE_BASELINE = {
    "removeElements": [
      {
        "namespace": "http://www.w3.org/1999/xhtml",
        "name": "script"
      },
      {
        "namespace": "http://www.w3.org/1999/xhtml",
        "name": "frame"
      },
      {
        "namespace": "http://www.w3.org/1999/xhtml",
        "name": "iframe"
      },
      {
        "namespace": "http://www.w3.org/1999/xhtml",
        "name": "object"
      },
      {
        "namespace": "http://www.w3.org/1999/xhtml",
        "name": "embed"
      },
      {
        "namespace": "http://www.w3.org/2000/svg",
        "name": "script"
      },
      {
        "namespace": "http://www.w3.org/2000/svg",
        "name": "use"
      }
    ],
    "removeAttributes": []
  };

  assert_equals(config.removeElements.length, SAFE_BASELINE.removeElements.length);
  for (let i = 0; i < SAFE_BASELINE.removeElements.length; i++) {
    let element = config.removeElements[i];
    assert_own_property(element, "name");
    assert_equals(element.name, SAFE_BASELINE.removeElements[i].name);
    assert_own_property(element, "namespace");
    assert_equals(element.namespace, SAFE_BASELINE.removeElements[i].namespace);
  }

  // This list depends on the implementation defined "event handler content attributes"
  assert_true(config.removeAttributes.length > 0, "Has removeAttributes");
  for (let attribute of config.removeAttributes) {
    assert_own_property(attribute, "name");
    assert_true(attribute.name.startsWith("on"), `attribute '${attribute.name}' starts with "on"`);
    assert_own_property(attribute, "namespace"); // XXX Maybe optional?
    assert_equals(attribute.namespace, null, "attribute is in null namespace");
  }
}, "removeUnsafe removes the right elements and attributes");

test(t => {
  let s = new Sanitizer("default");
  let before = s.get();

  let s2 = new Sanitizer("default");
  s2.removeUnsafe();
  let after = s2.get();

  // None of the default config elements are unsafe.
  assert_true(before.elements.length > 0);
  assert_equals(before.elements.length, after.elements.length, "elements don't change");

  // None of the default config attributes are unsafe.
  assert_true(before.attributes.length > 0);
  assert_equals(before.attributes.length, after.attributes.length, "attributes don't change");

  // Not in default config.
  assert_false('replaceWithChildrenElements' in before);
  assert_false('replaceWithChildrenElements' in after);
  assert_false('removeElements' in before);
  assert_false('removeElements' in after);
  assert_false('removeAttributes' in before);
  assert_false('removeAttributes' in after);
}, "removeUnsafe with default config")
</script>
</body>
</html>
